requirements.txt
absl-py==0.8.1
astor==0.8.0
attrs==19.3.0
chardet==3.0.4
cycler==0.10.0
dill==0.3.1.1
future==0.18.0
gast==0.2.2
google-pasta==0.1.7
googleapis-common-protos==1.6.0
grpcio==1.24.1
h5py==2.10.0
idna==2.8
javalang==0.12.0
joblib==0.14.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
numpy==1.17.2
opt-einsum==3.1.0
pandas==0.25.1
promise==2.2.1
protobuf==3.10.0
psutil==5.6.3
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
tensorboard==2.0.0
tensorflow-datasets==1.2.0
tensorflow-estimator==2.0.0
tensorflow-gpu==2.0.0
tensorflow-metadata==0.15.0
termcolor==1.1.0
tqdm==4.36.1
urllib3==1.25.6
Werkzeug==0.16.0

